requirements.txt
google-generativeai
supabase
fastapi
uvicorn
openai